requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.10
matplotlib>=3.7
plotly>=5.14
typer>=0.9.0
fastapi>=0.100
uvicorn>=0.22
streamlit>=1.25
yfinance>=0.2
pandas-ta>=0.3
pydantic>=1.10
pyyaml>=6.0
jinja2>=3.1
joblib>=1.3
rich>=13.0
pytest>=7.4
